chore(config): Add support for `media.tenor.com` and `media1.tenor.com` domain for next image
This commit updates the next.config file to include the hostnames "media.tenor.com" and "media1.tenor.com" as remote patterns. This allows the application to fetch media files from these hostnames using the HTTPS protocol. This change enables the application to retrieve media files from Tenor for use in the application.