feat: Add initial data for post route
This commit adds the initial data for the post route. It includes a new file `data.ts` which exports an array of post data. Each post object contains properties such as id, userId, content, likesCount, and media. This data will be used to populate the post route in the application.